Refatora is_boosted_day para módulo em functions
