+++ bot/constants/item.py
@@ -24,26 +24,6 @@ ESCAPED_CALLBACK_TEXT_GET = re.escape(CALLBACK_TEXT_GET)
 ESCAPED_CALLBACK_TEXT_IGNORE = re.escape(CALLBACK_TEXT_IGNORE)
 
 
-BOOSTED_DAYS = [
-    date(year=2000, month=1, day=1),  # Confraternização Universal
-    date(year=2000, month=4, day=21),  # Tiradentes
-    date(year=2000, month=5, day=1),  # Dia do Trabalho
-    date(year=2000, month=9, day=7),  # Independência do Brasil
-    date(year=2000, month=10, day=12),  # Nossa Senhora Aparecida (Crianças)
-    date(year=2000, month=11, day=2),  # Finados
-    date(year=2000, month=11, day=15),  # Proclamação da República
-    date(year=2000, month=12, day=8),  # Nossa Senhora da Imaculada Conceição
-    date(year=2000, month=12, day=24),  # Véspera de Natal
-    date(year=2000, month=12, day=25),  # Natal
-    date(year=2000, month=12, day=26),  # Recesso
-    date(year=2000, month=12, day=27),  # Recesso
-    date(year=2000, month=12, day=28),  # Recesso
-    date(year=2000, month=12, day=29),  # Recesso
-    date(year=2000, month=12, day=30),  # Recesso
-    date(year=2000, month=12, day=31),  # Véspera de Ano Novo
-]
-
-
 TRAP_TYPE_DAMAGE_MULTIPLIER = {
     DamageEnum.BLUDGEONING.name: 1.5,
     DamageEnum.HITTING.name: 1.5,
